+++ CHANGES.md
@@ -3,5 +3,6 @@
 * $Id$
 
 #17.0.0 (2017/xx/xx)
+ * (2016/12/23) - CR: T2016121990002948 (address address book functionality) (rbo)
  * (2016/12/06) - CR: T2016121190001552 (code merge of all packages) (ddoerffel)
  * first productive release